app/requirements.txt
torch
flask
numpy
pillow
torchvision
lmdb
tqdm
matplotlib
huggingface_hub
